Made quote command prettier and added permissions systen

+++ main.py
@@ -13,11 +13,17 @@ from helpers import read_file, replies, write_file
 import command_logic as logic
 from command_logic import meme, award, image, verse, quote, delete, cointoss, dice, wiki
 
-
 # Constants -------------------------------------------------------
 
 DEF_COLOR = 0x0089a1
 RED_COLOR = 0x800000
+ORA_COLOR = 0xff5900
+
+ICON_ERROR = 'https://freeiconshop.com/wp-content/uploads/edd/error-flat.png'
+ICON_MEDAL = 'https://cdn.discordapp.com/attachments/724745384840396953/879030788723724398/olympic_medal1600.png'
+ICON_PERMS = 'https://icons.iconarchive.com/icons/paomedia/small-n-flat/1024/sign-error-icon.png'
+ICON_BIBLE = 'https://cdn3.vectorstock.com/i/1000x1000/06/57/bible-icon-vector-5020657.jpg'
+
 PREFIX = '!'
 
 # Credentials -----------------------------------------------------
@@ -34,9 +40,10 @@ intents = discord.Intents().all()
 client = commands.Bot(command_prefix=PREFIX, intents=intents)
 slash = SlashCommand(client, sync_commands=True)
 
-commands = read_file.read_command_config()
+cmds = read_file.read_command_config()
 guilds = []
 author = ""
+
 
 # -----------------------------------------------------------------
 
@@ -63,6 +70,34 @@ def get_name(member: discord.member):
     else:
         return member.name
 
+
+async def no_perms(ctx):
+    perms = cmds[ctx.invoked_with]['permissions']
+    if perms and ctx.author.id not in perms:
+        await send_unauthorised(ctx)
+        return True
+
+    return False
+
+
+async def send_unauthorised(ctx):
+    try:
+        message = replies.get_reply(cmds[ctx.invoked_with]['unauthorised'])
+    except KeyError:
+        message = 'Dafür hast du leider nicht die nötigen Berechtigungen.'
+
+    embed_var = discord.Embed(
+        color=discord.Colour(RED_COLOR),
+        description=message
+    )
+    embed_var.set_author(
+        name='Keine Berechtigung',
+        icon_url=ICON_PERMS
+    )
+
+    await ctx.send(embed=embed_var)
+
+
 # -----------------------------------------------------------------
 
 
@@ -74,6 +109,7 @@ async def on_ready():
         write_file.write_guild_config(guild, ["logs", "temp", "user"])
 
     print(f'I am logged in ({client.user}) and ready!')
+
 
 # -----------------------------------------------------------------
 
@@ -94,6 +130,7 @@ async def on_message(message: discord.Message):
 
     await client.process_commands(message)
 
+
 # Slash commands --------------------------------------------------
 
 
@@ -103,9 +140,7 @@ async def on_message(message: discord.Message):
     guild_ids=guilds
 )
 async def _hello(ctx):
-    message = replies.enrich(
-        replies.get_reply(commands[ctx.name]['replies']), mention=ctx.author.mention)
-    await ctx.send(message)
+    await hello(ctx)
 
 
 @slash.slash(
@@ -225,25 +260,31 @@ async def _award(ctx, mention=None):
             required=True,
             option_type=3,
             choices=[
-                create_choice(name="Hilfe", value="help"),
-                create_choice(name="Hallo", value="hello"),
-                create_choice(name="Würfel", value="dice"),
+                create_choice(name="Help", value="help"),
+                create_choice(name="Hello", value="hello"),
+                create_choice(name="Dice", value="dice"),
                 create_choice(name="Münzwurf", value="cointoss"),
                 create_choice(name="Person", value="person"),
-                create_choice(name="Katze", value="cat"),
+                create_choice(name="Cat", value="cat"),
                 create_choice(name="Award", value="award"),
                 create_choice(name="Awards", value="awards"),
                 create_choice(name="Essen", value="eating"),
                 create_choice(name="Essensstats", value="eatingstats"),
                 create_choice(name="Vers", value="verse"),
-                create_choice(name="Löschen", value="delete")
+                create_choice(name="Löschen", value="delete"),
+                create_choice(name="Wiki", value="wiki"),
+                create_choice(name="Offtopic", value="offtopic"),
+                create_choice(name="Morgen", value="morning"),
+                create_choice(name="Anruf", value="call"),
+                create_choice(name="Langeweile", value="bored"),
+                create_choice(name="Verlassen", value="quit"),
             ]
         )
     ]
 )
 async def _help(ctx, command=None):
     try:
-        message = commands[command].help
+        message = cmds[command].help
     except KeyError:
         message = 'Es existiert keine Hilfe für diesen Befehl'
 
@@ -254,32 +295,88 @@ async def _help(ctx, command=None):
 
 
 @client.command()
-async def dice(ctx):
-    message = replies.enrich(
-        replies.get_reply(commands[ctx.invoked_with]['replies']),
-        insert=logic.dice.dice_toss(),
+async def hello(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
+    message = replies.enrich(replies.get_reply(
+        cmds[ctx.invoked_with]['replies']),
         mention=ctx.author.mention
     )
-    await ctx.send(message)
+
+    # Create embed
+    embed_var = discord.Embed(
+        title='Hallo',
+        color=discord.Colour(DEF_COLOR),
+        description=message)
+
+    await ctx.send(embed=embed_var)
 
 
 @client.command()
-async def wiki(ctx, arg):
+async def quit(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
+    message = replies.enrich(replies.get_reply(cmds[ctx.invoked_with]['replies']))
+
+    # Create embed
+    embed_var = discord.Embed(
+        title='Quit',
+        color=discord.Colour(DEF_COLOR),
+        description=message)
+
+    await ctx.send(embed=embed_var)
+    await client.close()
+    exit(0)
+
+
+@client.command()
+async def dice(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
+    message = replies.enrich(
+        replies.get_reply(cmds[ctx.invoked_with]['replies']),
+        insert=logic.dice.dice_toss(),
+        mention=ctx.author.mention
+    )
+
+    # Create embed
+    embed_var = discord.Embed(
+        title='Dice',
+        color=discord.Colour(DEF_COLOR),
+        description=message)
+
+    await ctx.send(embed=embed_var)
+
+
+@client.command()
+async def wiki(ctx, arg=None):
+    denied = await no_perms(ctx)
+    if denied: return
+
     # Test for params
     if not arg:
         embed_var = discord.Embed(
-            title='Error',
             color=discord.Colour(RED_COLOR),
-            description=replies.get_reply(commands[ctx.invoked_with]['error'])
-            )
+            description=replies.get_reply(cmds[ctx.invoked_with]['error'])
+        )
+        embed_var.set_author(
+            name='Error',
+            icon_url=ICON_ERROR
+        )
 
-    link = logic.wiki.get_wiki(commands[ctx.invoked_with]['urls'][0], arg)
+        await ctx.send(embed=embed_var)
+        return
+
+    link = logic.wiki.get_wiki(cmds[ctx.invoked_with]['urls'][0], arg)
 
     # Get reply depending on if the site exists
     if 'search' in link:
-        message = replies.enrich(replies.get_reply(commands[ctx.invoked_with]['replies2']))
+        message = replies.enrich(replies.get_reply(cmds[ctx.invoked_with]['replies2']))
     else:
-        message = replies.enrich(replies.get_reply(commands[ctx.invoked_with]['replies']))
+        message = replies.enrich(replies.get_reply(cmds[ctx.invoked_with]['replies']))
 
     # Create embed
     embed_var = discord.Embed(
@@ -295,18 +392,25 @@ async def wiki(ctx, arg):
 
 @client.command()
 async def award(ctx, *args):
-    if len(args) != 1 or not (args[0].startswith('<@!') and args[0].endswith('>')):
-        await ctx.send('Error')  # TODO: Read reply from bot.json
-        return
+    denied = await no_perms(ctx)
+    if denied: return
 
-    medal = "https://cdn.discordapp.com/attachments/724745384840396953/879030788723724398/olympic_medal1600.png"
+    if len(args) != 1 or not (args[0].startswith('<@!') and args[0].endswith('>')):
+        embed_var = discord.Embed(
+            title='Error',
+            color=discord.Colour(RED_COLOR),
+            description=replies.get_reply(cmds[ctx.invoked_with]['error'])
+        )
+
+        await ctx.send(embed=embed_var)
+        return
 
     # Give award
     logic.award.give_award(ctx.channel.guild.id, args[0][3:-1])
 
     # Get message
     message = replies.enrich(
-        replies.get_reply(commands[ctx.invoked_with]['replies']),
+        replies.get_reply(cmds[ctx.invoked_with]['replies']),
         ref_mention=args[0],
         mention=ctx.author.mention
     )
@@ -316,13 +420,16 @@ async def award(ctx, *args):
         color=discord.Colour(DEF_COLOR),
         description=message
     )
-    embed_var.set_author(name=f"Award von {get_name(ctx.author)}", icon_url=medal)
+    embed_var.set_author(name=f"Award von {get_name(ctx.author)}", icon_url=ICON_MEDAL)
 
     await ctx.send(embed=embed_var)
 
 
 @client.command()
 async def meme(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
     meme_object = logic.meme.get_meme(reddit, ctx.channel.guild.id)
     embed_var = discord.Embed(title='Meme', color=discord.Colour(DEF_COLOR))
     embed_var.set_image(url=meme_object.url)
@@ -336,36 +443,86 @@ async def meme(ctx):
 
 @client.command()
 async def quote(ctx):
-    message = logic.quote.get_quote(commands[ctx.invoked_with]['urls'][0])
-    embed_var = discord.Embed(title='Zitat', color=discord.Colour(DEF_COLOR))
-    embed_var.add_field(name=message[1], value=message[0], inline=True)
+    denied = await no_perms(ctx)
+    if denied: return
+
+    message = logic.quote.get_quote(cmds[ctx.invoked_with]['urls'][0])
+    embed_var = discord.Embed(
+        description=message['quote'],
+        color=discord.Colour(DEF_COLOR))
+    embed_var.set_author(
+        name=message['author'],
+        icon_url=message['image']
+    )
+
+    await ctx.send(embed=embed_var)
+
+
+@client.command()
+async def ping(ctx):
+    if no_perms(ctx):
+        return
+
+    bot_ping = "{:.2f}".format(client.latency * 100)
+    reply = replies.enrich(replies.get_reply(ctx.invoked_with['replies']), insert=bot_ping + ' ms')
+    embed_var = discord.Embed(title='Ping', colour=discord.Colour(DEF_COLOR), description=reply)
     await ctx.send(embed=embed_var)
 
 
 @client.command()
 async def verse(ctx):
-    message = logic.verse.get_verse(commands[ctx.invoked_with]['urls'][0])
-    embed_var = discord.Embed(title='Bibelvers', color=discord.Colour(DEF_COLOR))
+    denied = await no_perms(ctx)
+    if denied: return
+
+    message = logic.verse.get_verse(cmds[ctx.invoked_with]['urls'][0])
+    embed_var = discord.Embed(color=discord.Colour(DEF_COLOR))
     embed_var.add_field(name=message[1], value=message[0], inline=True)
+    embed_var.set_author(
+        name="Bibel",
+        icon_url=ICON_BIBLE
+    )
     await ctx.send(embed=embed_var)
 
 
 @client.command()
 async def cointoss(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
     message = replies.enrich(
-        replies.get_reply(commands[ctx.invoked_with]['replies']),
+        replies.get_reply(cmds[ctx.invoked_with]['replies']),
         insert=logic.cointoss.coin_toss(),
         mention=ctx.author.mention)
     await ctx.send(message)
 
 
 @client.command()
+async def commands(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
+    message = replies.get_reply(cmds[ctx.invoked_with]['replies'])
+    embed_var = discord.Embed(
+        color=discord.Colour(ORA_COLOR),
+        description=message
+    )
+    embed_var.set_author(
+        name='Error',
+        icon_url=ICON_ERROR
+    )
+    await ctx.send(embed=embed_var)
+
+
+@client.command()
 async def cat(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
     # Get the image
     file = logic.image.from_url(f'guilds/{ctx.guild.id}/temp/cat.png',
-                                replies.get_reply(commands[ctx.invoked_with]['urls']))
+                                replies.get_reply(cmds[ctx.invoked_with]['urls']))
 
-    message = replies.get_reply(commands[ctx.invoked_with]['replies'])
+    message = replies.get_reply(cmds[ctx.invoked_with]['replies'])
     embed_var = discord.Embed(title='Cat', color=discord.Colour(DEF_COLOR), description=message)
     embed_var.set_image(url="attachment://image.png")
     await ctx.send(file=file, embed=embed_var)
@@ -373,11 +530,14 @@ async def cat(ctx):
 
 @client.command()
 async def person(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
     # Get the image
     file = logic.image.from_url(f'guilds/{ctx.guild.id}/temp/person.png',
-                                replies.get_reply(commands[ctx.invoked_with]['urls']))
+                                replies.get_reply(cmds[ctx.invoked_with]['urls']))
 
-    message = replies.get_reply(commands[ctx.invoked_with]['replies'])
+    message = replies.get_reply(cmds[ctx.invoked_with]['replies'])
     embed_var = discord.Embed(title='Person', color=discord.Colour(DEF_COLOR), description=message)
     embed_var.set_image(url=f"attachment://image.png")
     await ctx.send(file=file, embed=embed_var)
@@ -385,6 +545,22 @@ async def person(ctx):
 
 @client.command()
 async def delete(ctx):
+    denied = await no_perms(ctx)
+    if denied: return
+
+    if not ctx.message.reference:
+        embed_var = discord.Embed(
+            color=discord.Colour(RED_COLOR),
+            description=replies.get_reply(cmds[ctx.invoked_with]['error'])
+        )
+        embed_var.set_author(
+            name='Error',
+            icon_url=ICON_ERROR
+        )
+
+        await ctx.send(embed=embed_var)
+        return
+
     await logic.delete.delete_message(ctx)
